refactor: (BIG) refactor file-picker, separate the search & preview logic
